Rename chain to chain_of_responsibility
