+++ patterns/behavioral/chain_of_responsibility.py
